requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
transformers==4.32.0
torch==2.1.0
sentence-transformers==2.2.2
pdfplumber  #==0.7.8
python-docx==0.8.11
pymilvus==2.0.0
faiss-cpu==1.7.4
numpy==1.24.0
scikit-learn==1.2.1
pytest==7.2.2
mlflow==2.6.0